+++ CHANGELOG.md
@@ -8,6 +8,7 @@ ENHANCEMENTS:
 * resource/spotinst_ocean_aws: added `auto_headroom_percentage`
 * resource/spotinst_ocean_aws_launch_spec: added `tags`
 * resource/spotinst_ocean_aws: added `grace_period`
+* resource/spotinst_elastigroup_aws: added `OnFailure` under `roll_config`
 
 NOTES: 
 * documentation: resource/spotinst_mrscaler_aws: retries are now accurate